// src/user/user.type.ts
export class GoogleUserInfo {
  sub: string;
  name: string;
  picture: 'https://lh3.googleusercontent.com/a/AGNmyxYMFJjbjccLF9olEbK5V6lGmYZKt41xrXKx_t5I=s96-c';
  email: string;
}

export class UserInfo {
  id: string;
  name: string;
  picture: 'https://lh3.googleusercontent.com/a/AGNmyxYMFJjbjccLF9olEbK5V6lGmYZKt41xrXKx_t5I=s96-c';
  email: string;
  key: string;
}